+++ CHANGELOG.md
@@ -7,9 +7,10 @@
 # CHANGELOG
 
 <p>
-<b>Version 2.2.***************</b> - October 2023<br/>
+<b>Version 2.2.2</b> - October 2023<br/>
 - NEW: HUE Motion: support HUE Camera motion events via the HUE Motion node.<br/>
 - HUE Light: some tweaking to the GUI.<br/> 
+- HUE Grouped Light: Fixed relative dimming function.<br/> 
 - KNK Alerter node: <a href="https://github.com/Supergiovane/node-red-contrib-knx-ultimate/wiki/SampleAlerter">fixed the wiki sample page</a>.<br/> 
 - KNX Viewer: added the help pane in Node-Red.<br/> 
 </p>